folder restructure and added route for calendar
