+++ src/Components/AppNavbar/AppNavbar.js
@@ -4,8 +4,8 @@ import NavUserIcon from './components/NavUserIcon';
 import Logo from '../Subcomponents/Logo';
 import UserIconPop from './components/UserIconPop';
 import { useLocation } from 'react-router-dom';
-import UserContext from '../../userContext';
-import NotificationContext from '../../notificationContext';
+import UserContext from '../../Contexts/userContext';
+import NotificationContext from '../../Contexts/notificationContext';
 import NavNotificationIcon from './components/NavNotificationIcon';
 import io from "socket.io-client";
 import NavNotificationPop from './components/NavNotificationPop';
