+++ src/Pages/Logout/Logout.js
@@ -1,6 +1,6 @@
 import { useContext, useEffect } from 'react';
 import { Navigate } from 'react-router-dom';
-import UserContext from '../../userContext';
+import UserContext from '../../Contexts/userContext';
 
 export default function Logout() {
