+++ src/Pages/MyCalendar/MyCalendar.js
@@ -1,0 +1,13 @@
+
+export default function MyCalendar(){
+
+    return(
+        <>
+            <div className="p-3 mt-3" style={{borderStyle:'solid', borderColor:'#516473', borderRadius:'8px', borderWidth:'1px'}}>
+            <h6 className="text-muted">You are in the MyCalendar Page. Please work on me. Kindly refer to this figma link to see how this page is supposed to look like.</h6>
+            <a href="https://www.figma.com/file/yL80hD5gDmu7QpaXHln50L/HSQ-PORTAL?type=design&node-id=0%3A1&mode=design&t=UjTq5WUeOD8XRB1Q-1" target="_blank">Figma link</a>
+            </div>
+            
+        </>
+    )
+}
